+++ app/src/main/java/org/simple/clinic/signature/SignatureEvent.kt
@@ -1,0 +1,13 @@
+package org.simple.clinic.signature
+
+import android.graphics.Bitmap
+
+sealed class SignatureEvent
+
+object UndoClicked : SignatureEvent()
+
+data class AcceptClicked(
+    val bitmap: Bitmap?
+) : SignatureEvent()
+
+object SignatureAccepted : SignatureEvent()