+++ app/src/main/java/org/simple/clinic/signature/SignatureModel.kt
@@ -1,0 +1,17 @@
+package org.simple.clinic.signature
+
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+import java.io.File
+
+@Parcelize
+data class SignatureModel(
+    val internalStoragePath: File
+) : Parcelable {
+
+  companion object {
+    fun create(path: File) = SignatureModel(
+        internalStoragePath = path
+    )
+  }
+}